fix: 취소를 눌렀을시 삭제되는 오류 해결
